+++ index.js
@@ -1,16 +1,17 @@
-'use strict'
+"use strict";
 
-const path = require('path')
-const fs = require('fs')
-const globby = require('globby')
-const crypto = require('crypto')
-const { CleanWebpackPlugin } = require('clean-webpack-plugin')
-const WorkboxPlugin = require('workbox-webpack-plugin')
-const defaultCache = require('./cache')
-const buildCustomWorker = require('./build-custom-worker')
-const buildFallbackWorker = require('./build-fallback-worker')
+const path = require("path");
+const fs = require("fs");
+const globby = require("globby");
+const crypto = require("crypto");
+const { CleanWebpackPlugin } = require("clean-webpack-plugin");
+const WorkboxPlugin = require("workbox-webpack-plugin");
+const defaultCache = require("./cache");
+const buildCustomWorker = require("./build-custom-worker");
+const buildFallbackWorker = require("./build-fallback-worker");
 
-const getRevision = file => crypto.createHash('md5').update(fs.readFileSync(file)).digest('hex')
+const getRevision = (file) =>
+  crypto.createHash("md5").update(fs.readFileSync(file)).digest("hex");
 
 module.exports = (nextConfig = {}) => ({
   ...nextConfig,
@@ -19,11 +20,16 @@ module.exports = (nextConfig = {}) => ({
       webpack,
       buildId,
       dev,
-      config: { distDir = '.next', pwa = {}, experimental = {}}
-    } = options
+      config: {
+        distDir = ".next",
+        pwa = {},
+        pageExtensions = ["tsx", "ts", "jsx", "js", "mdx"],
+        experimental = {},
+      },
+    } = options;
 
-    let basePath = options.config.basePath
-    if (!basePath) basePath = '/'
+    let basePath = options.config.basePath;
+    if (!basePath) basePath = "/";
 
     // For workbox configurations:
     // https://developers.google.com/web/tools/workbox/reference-docs/latest/module-workbox-webpack-plugin.GenerateSW
@@ -31,7 +37,8 @@ module.exports = (nextConfig = {}) => ({
       disable = false,
       register = true,
       dest = distDir,
-      sw = 'sw.js',
+      sw = "sw.js",
+      cacheStartUrl = true,
       dynamicStartUrl = true,
       dynamicStartUrlRedirect,
       skipWaiting = true,
@@ -40,248 +47,306 @@ module.exports = (nextConfig = {}) => ({
       additionalManifestEntries,
       ignoreURLParametersMatching = [],
       importScripts = [],
-      publicExcludes = ['!noprecache/**/*'],
+      publicExcludes = ["!noprecache/**/*"],
       buildExcludes = [],
-      manifestTransforms = [],
       modifyURLPrefix = {},
+      manifestTransforms = [],
       fallbacks = {},
       cacheOnFrontEndNav = false,
+      reloadOnOnline = true,
       scope = basePath,
-      subdomainPrefix,  // deprecated, use basePath in next.config.js instead
-      staticDomainPrefix  = '',
+      staticDomainPrefix = "",
+      customWorkerDir = "worker",
+      subdomainPrefix, // deprecated, use basePath in next.config.js instead
       ...workbox
-    } = pwa
+    } = pwa;
 
-    if (typeof nextConfig.webpack === 'function') {
-      config = nextConfig.webpack(config, options)
+    if (typeof nextConfig.webpack === "function") {
+      config = nextConfig.webpack(config, options);
     }
 
     if (disable) {
-      options.isServer && console.log('> [PWA] PWA support is disabled')
-      return config
+      options.isServer && console.log("> [PWA] PWA support is disabled");
+      return config;
     }
 
     if (subdomainPrefix) {
-      console.error('> [PWA] subdomainPrefix is deprecated, use basePath in next.config.js instead: https://nextjs.org/docs/api-reference/next.config.js/basepath')
+      console.error(
+        "> [PWA] subdomainPrefix is deprecated, use basePath in next.config.js instead: https://nextjs.org/docs/api-reference/next.config.js/basepath"
+      );
     }
 
-    console.log(`> [PWA] Compile ${options.isServer ? 'server' : 'client (static)'}`)
-    
-    let { runtimeCaching = defaultCache } = pwa
-    const _scope = path.posix.join(scope, '/')
+    console.log(
+      `> [PWA] Compile ${options.isServer ? "server" : "client (static)"}`
+    );
+
+    let { runtimeCaching = defaultCache } = pwa;
+    const _scope = path.posix.join(scope, "/");
 
     // inject register script to main.js
-    const _sw = path.posix.join(basePath, sw.startsWith('/') ? sw : `/${sw}`)
+    const _sw = path.posix.join(basePath, sw.startsWith("/") ? sw : `/${sw}`);
     config.plugins.push(
       new webpack.DefinePlugin({
         __PWA_SW__: `'${_sw}'`,
         __PWA_SCOPE__: `'${_scope}'`,
         __PWA_ENABLE_REGISTER__: `${Boolean(register)}`,
         __PWA_START_URL__: dynamicStartUrl ? `'${basePath}'` : undefined,
-        __PWA_CACHE_ON_FRONT_END_NAV__: `${Boolean(cacheOnFrontEndNav)}`
+        __PWA_CACHE_ON_FRONT_END_NAV__: `${Boolean(cacheOnFrontEndNav)}`,
+        __PWA_RELOAD_ON_ONLINE__: `${Boolean(reloadOnOnline)}`,
       })
-    )
+    );
 
-    const registerJs = path.join(__dirname, 'register.js')
-    const entry = config.entry
+    const registerJs = path.join(__dirname, "register.js");
+    const entry = config.entry;
     config.entry = () =>
-      entry().then(entries => {
-        if (entries['main.js'] && !entries['main.js'].includes(registerJs)) {
-          entries['main.js'].unshift(registerJs)
+      entry().then((entries) => {
+        if (entries["main.js"] && !entries["main.js"].includes(registerJs)) {
+          entries["main.js"].unshift(registerJs);
         }
-        return entries
-      })
+        return entries;
+      });
 
     if (!options.isServer) {
-      const _dest = path.join(options.dir, dest)
+      const _dest = path.join(options.dir, dest);
       buildCustomWorker({
         id: buildId,
         basedir: options.dir,
+        customWorkerDir,
         destdir: _dest,
-        plugins: config.plugins.filter(plugin => plugin instanceof webpack.DefinePlugin),
-        success: ({name}) => importScripts.unshift(name),
-        minify: !dev
-      })
+        plugins: config.plugins.filter(
+          (plugin) => plugin instanceof webpack.DefinePlugin
+        ),
+        success: ({ name }) => importScripts.unshift(name),
+        minify: !dev,
+      });
 
       if (register) {
-        console.log(`> [PWA] Auto register service worker with: ${path.resolve(registerJs)}`)
+        console.log(
+          `> [PWA] Auto register service worker with: ${path.resolve(
+            registerJs
+          )}`
+        );
       } else {
-        console.log(`> [PWA] Auto register service worker is disabled, please call following code in componentDidMount callback or useEffect hook`)
-        console.log(`> [PWA]   window.workbox.register()`)
+        console.log(
+          `> [PWA] Auto register service worker is disabled, please call following code in componentDidMount callback or useEffect hook`
+        );
+        console.log(`> [PWA]   window.workbox.register()`);
       }
 
-      console.log(`> [PWA] Service worker: ${path.join(_dest, sw)}`)
-      console.log(`> [PWA]   url: ${_sw}`)
-      console.log(`> [PWA]   scope: ${_scope}`)
+      console.log(`> [PWA] Service worker: ${path.join(_dest, sw)}`);
+      console.log(`> [PWA]   url: ${_sw}`);
+      console.log(`> [PWA]   scope: ${_scope}`);
 
       config.plugins.push(
         new CleanWebpackPlugin({
           cleanOnceBeforeBuildPatterns: [
-            path.join(_dest, 'workbox-*.js'),
-            path.join(_dest, 'workbox-*.js.map'),
+            path.join(_dest, "workbox-*.js"),
+            path.join(_dest, "workbox-*.js.map"),
             path.join(_dest, sw),
-            path.join(_dest, `${sw}.map`)
-          ]
+            path.join(_dest, `${sw}.map`),
+          ],
         })
-      )
+      );
 
       // precache files in public folder
-      let manifestEntries = additionalManifestEntries
+      let manifestEntries = additionalManifestEntries;
       if (!Array.isArray(manifestEntries)) {
         manifestEntries = globby
           .sync(
             [
-              '**/*',
-              '!workbox-*.js',
-              '!workbox-*.js.map',
-              '!worker-*.js',
-              '!worker-*.js.map',
-              '!fallback-*.js',
-              '!fallback-*.js.map',
-              `!${sw.replace(/^\/+/, '')}`,
-              `!${sw.replace(/^\/+/, '')}.map`,
-              ...publicExcludes
+              "**/*",
+              "!workbox-*.js",
+              "!workbox-*.js.map",
+              "!worker-*.js",
+              "!worker-*.js.map",
+              "!fallback-*.js",
+              "!fallback-*.js.map",
+              `!${sw.replace(/^\/+/, "")}`,
+              `!${sw.replace(/^\/+/, "")}.map`,
+              ...publicExcludes,
             ],
             {
-              cwd: 'public'
+              cwd: "public",
             }
           )
-          .map(f => ({
+          .map((f) => ({
             url: staticDomainPrefix + path.posix.join(basePath, `/${f}`),
-            revision: getRevision(`public/${f}`)
-          }))
+            revision: getRevision(`public/${f}`),
+          }));
       }
 
-      if (!dynamicStartUrl) {
-        manifestEntries.push({
-          url: basePath,
-          revision: buildId
-        })
-      } else if (typeof dynamicStartUrlRedirect === 'string' && dynamicStartUrlRedirect.length > 0) {
-        manifestEntries.push({
-          url: dynamicStartUrlRedirect,
-          revision: buildId
-        })
+      if (cacheStartUrl) {
+        if (!dynamicStartUrl) {
+          manifestEntries.push({
+            url: basePath,
+            revision: buildId,
+          });
+        } else if (
+          typeof dynamicStartUrlRedirect === "string" &&
+          dynamicStartUrlRedirect.length > 0
+        ) {
+          manifestEntries.push({
+            url: dynamicStartUrlRedirect,
+            revision: buildId,
+          });
+        }
       }
 
-      let _fallbacks = fallbacks
+      let _fallbacks = fallbacks;
       if (_fallbacks) {
         _fallbacks = buildFallbackWorker({
           id: buildId,
           fallbacks,
           basedir: options.dir,
           destdir: _dest,
-          success: ({name, precaches}) => {
-            importScripts.unshift(name)
-            precaches.forEach(route => {
-              if (!manifestEntries.find(entry => entry.url.startsWith(route))) {
+          success: ({ name, precaches }) => {
+            importScripts.unshift(name);
+            precaches.forEach((route) => {
+              if (
+                !manifestEntries.find((entry) => entry.url.startsWith(route))
+              ) {
                 manifestEntries.push({
                   url: route,
-                  revision: buildId
-                })
+                  revision: buildId,
+                });
               }
-            })
+            });
           },
-          minify: !dev
-        })
+          minify: !dev,
+          pageExtensions,
+        });
       }
 
-      const prefix = config.output.publicPath ? `${config.output.publicPath}static/` : 'static/'
       const workboxCommon = {
         swDest: path.join(_dest, sw),
         additionalManifestEntries: dev ? [] : manifestEntries,
         exclude: [
+          ...buildExcludes,
           ({ asset, compilation }) => {
-            if (asset.name.match(/^(build-manifest\.json|react-loadable-manifest\.json)$/)) {
-              return true
+            if (
+              asset.name.startsWith("server/") ||
+              asset.name.match(
+                /^(build-manifest\.json|react-loadable-manifest\.json)$/
+              )
+            ) {
+              return true;
             }
-            if (dev && !asset.name.startsWith('static/runtime/')) {
-              return true
+            if (dev && !asset.name.startsWith("static/runtime/")) {
+              return true;
             }
             if (experimental.modern /* modern */) {
-              if (asset.name.endsWith('.module.js')) {
-                return false
+              if (asset.name.endsWith(".module.js")) {
+                return false;
               }
-              if (asset.name.endsWith('.js')) {
-                return true
+              if (asset.name.endsWith(".js")) {
+                return true;
               }
             }
-            return false
+            return false;
           },
-          ...buildExcludes
         ],
         modifyURLPrefix: {
           ...modifyURLPrefix,
-          [prefix]: staticDomainPrefix + path.posix.join(basePath, '/_next/static/')
+          "/_next/../public/": staticDomainPrefix + "/",
         },
         manifestTransforms: [
           ...manifestTransforms,
           async (manifestEntries, compilation) => {
-            const manifest = manifestEntries.map(m => {
-              m.url = m.url.replace(/\/\[/g, '/%5B').replace(/\]/g, '%5D')
-              m.revision = buildId
-              return m
-            })
-            return { manifest, warnings: [] }
-          }
-        ]
-      }
+            const manifest = manifestEntries.map((m) => {
+              m.url = m.url.replace(
+                "/_next//static/image",
+                "/_next/static/image"
+              );
+              m.url = m.url.replace(
+                "/_next//static/media",
+                "/_next/static/media"
+              );
+              if (m.revision === null) {
+                let key = m.url;
+                if (key.startsWith(config.output.publicPath)) {
+                  key = m.url.substring(config.output.publicPath.length);
+                }
+                const assset = compilation.assetsInfo.get(key);
+                m.revision = assset ? assset.contenthash || buildId : buildId;
+              }
+              m.url = m.url.replace(/\[/g, "%5B").replace(/\]/g, "%5D");
+              return m;
+            });
+            return { manifest, warnings: [] };
+          },
+        ],
+      };
 
       if (workbox.swSrc) {
-        const swSrc = path.join(options.dir, workbox.swSrc)
-        console.log(`> [PWA] Inject manifest in ${swSrc}`)
+        const swSrc = path.join(options.dir, workbox.swSrc);
+        console.log(`> [PWA] Inject manifest in ${swSrc}`);
         config.plugins.push(
           new WorkboxPlugin.InjectManifest({
             ...workboxCommon,
             ...workbox,
-            swSrc
+            swSrc,
           })
-        )
+        );
       } else {
         if (dev) {
           console.log(
-            '> [PWA] Build in develop mode, cache and precache are mostly disabled. This means offline support is disabled, but you can continue developing other functions in service worker.'
-          )
+            "> [PWA] Build in develop mode, cache and precache are mostly disabled. This means offline support is disabled, but you can continue developing other functions in service worker."
+          );
 
-          ignoreURLParametersMatching.push(/ts/)
+          ignoreURLParametersMatching.push(/ts/);
           runtimeCaching = [
             {
               urlPattern: /.*/i,
-              handler: 'NetworkOnly',
+              handler: "NetworkOnly",
               options: {
-                cacheName: 'dev'
-              }
-            }
-          ]
+                cacheName: "dev",
+              },
+            },
+          ];
         }
 
         if (dynamicStartUrl) {
           runtimeCaching.unshift({
             urlPattern: basePath,
-            handler: 'NetworkFirst',
+            handler: "NetworkFirst",
             options: {
-              cacheName: 'start-url',
-              plugins: [{
-                cacheWillUpdate: async ({request, response, event, state}) => {
-                  if (response && response.type === 'opaqueredirect') {
-                    return new Response(response.body, {status: 200, statusText: 'OK', headers: response.headers});
-                  }
-                  return response;
-                }
-              }]
-            }
-          })
+              cacheName: "start-url",
+              plugins: [
+                {
+                  cacheWillUpdate: async ({
+                    request,
+                    response,
+                    event,
+                    state,
+                  }) => {
+                    if (response && response.type === "opaqueredirect") {
+                      return new Response(response.body, {
+                        status: 200,
+                        statusText: "OK",
+                        headers: response.headers,
+                      });
+                    }
+                    return response;
+                  },
+                },
+              ],
+            },
+          });
         }
 
         if (_fallbacks) {
-          runtimeCaching.forEach(c => {
-            if (c.options.precacheFallback) return
-            if (Array.isArray(c.options.plugins) && c.options.plugins.find(p => 'handlerDidError' in p)) return
-            if (!c.options.plugins) c.options.plugins = []
+          runtimeCaching.forEach((c) => {
+            if (c.options.precacheFallback) return;
+            if (
+              Array.isArray(c.options.plugins) &&
+              c.options.plugins.find((p) => "handlerDidError" in p)
+            )
+              return;
+            if (!c.options.plugins) c.options.plugins = [];
             c.options.plugins.push({
-              handlerDidError: async ({request}) => self.fallback(request)
-            })
-          })
+              handlerDidError: async ({ request }) => self.fallback(request),
+            });
+          });
         }
 
         config.plugins.push(
@@ -293,12 +358,12 @@ module.exports = (nextConfig = {}) => ({
             ignoreURLParametersMatching,
             importScripts,
             ...workbox,
-            runtimeCaching
+            runtimeCaching,
           })
-        )
+        );
       }
     }
 
-    return config
-  }
-})
+    return config;
+  },
+});